fix(pathways): lead the row with the deadline, not the start date
The reviewed mockup led each row with "Starts 21 Oct". Checking the data
before merging: Course.cohortStartDate is null for every programme in
this catalogue. seed-pathway-programmes.ts writes sessionDates (free
text) and enrollByDate, and nothing writes a cohort start — so the
mockup's leading element would have rendered "Starts TBC" on all eleven
rows, which is worse than the panel it replaced.

The deadline takes the lead instead. It is populated, it is the date
that requires an action rather than describing one, and the closing-soon
state already keys off it — so the emphasised value and the urgent
colour are now the same thing rather than two competing dates.

startsLabel is still plumbed through and still renders when present, so
the row improves by itself if cohortStartDate is ever backfilled. That
backfill is worth doing: "when does it run" is a real question and
sessionDates only answers it as prose.

enrollByFmt drops to a short month — "Sep 11, 2026" rather than
"September 11, 2026" — now that it sets at 17px bold in a 12rem column.

+++ src/app/(dashboard)/pathways/page.tsx
@@ -177,7 +177,7 @@ export default async function PathwaysPage() {
   // Formatted server-side so a date cannot render differently after
   // hydration in another timezone.
   const enrollByFmt = new Intl.DateTimeFormat("en-CA", {
-    month: "long", day: "numeric", year: "numeric", timeZone: "America/Toronto",
+    month: "short", day: "numeric", year: "numeric", timeZone: "America/Toronto",
   });
   // The start date leads each programme row, so it is deliberately short —
   // "21 Oct" rather than the full date the deadline uses underneath it.
